inventory-service/app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
kafka-python==2.0.2
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.1
tenacity==8.2.3


uvicorn app.main:app --reload --port 8082